requirements.txt
google-generativeai
pytubefix
torch
numpy
transformers
moviepy